Added runlevel differences calculation.
The get_changed_init_scripts function allows to get a scripts lists without
previous runlevel scripts. This is done with a simple comparison between
the two directories that link to the scripts.

I have simplified the interface of serial_start and only need the runlevel
codes, to make this change I had to refactorize a bit of code.

With this change the execution of serial scripts is finished, although
some adjustments can be done :-O