Improvements in the selection of startup algorithms.
I have added a structure in which it is easier to select the startup
algorithm. It will only be necessary to add the corresponding code and
function for the new algorithm in startup_mode[].

Any error in the selection of the startup algorithm results in the use of the
serial start, it can be considered quite stable.

At the moment there are two methods to selec the start function, I am doubting
which one to use, they are very similar in structure, but the return value is
the key. Although it may be simple, I am thinking about how errors will be
handled in the future, and the use of one or the other would imply better
results in terms of code readability and performance. (Perhaps this right
now is unimportant U¬_¬)